Rename property functions to singular
The property-altering functions were renamed to be singular, rather
than plural. The naming was originally done to be symmetric with
CMake's naming convention of 'set_xxx_properties' -- however the naming
doesn't work quite as well for functions whose cardinality could be
referring to two different sources (either the source being copied, or
the destination).

The singular naming fits this better.